refactor: profile-image에 이름 렌더링 책임 이동하기
